+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/home/coordinator/CoordinatorHomeScreen.kt
@@ -17,6 +17,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -41,7 +42,7 @@ import com.app.arcabyolimpo.presentation.ui.components.molecules.FunctionalNavBa
 @Suppress("ktlint:standard:function-naming")
 @Composable
 fun CoordinatorHomeScreen(navController: NavHostController) {
-    var selectedTab by remember { mutableIntStateOf(0) }
+    var selectedTab by rememberSaveable { mutableIntStateOf(0) }
 
     Column(
         modifier = Modifier.fillMaxSize(),