+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchDetail/ProductBatchDetailScreen.kt
@@ -79,9 +79,6 @@ fun ProductBatchDetailScreen(
 
     Scaffold(
         containerColor = Background,
-        bottomBar = {
-            NavBar()
-        },
         topBar = {
             TopAppBar(
                 title = {